# openapi.yml
# openapi.yaml
openapi: 3.1.0
info:
  title: Restaurant Reservations API
  version: 1.0.0
  description: API for area/table management, slot-based availability, and reservations.
servers:
  - url: https://YOUR-RAILWAY-URL
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  time: { type: string, example: "2026-02-18T20:10:00-04:00" }

  /seed:
    post:
      summary: Seed initial data (optional but recommended)
      responses:
        "200":
          description: Seeded
        "409":
          description: Already seeded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: Seed file missing or invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /areas:
    get:
      summary: List areas
      responses:
        "200":
          description: Areas
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Area" }
        "503":
          description: Not seeded (run POST /seed first)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /areas/{areaId}/tables:
    post:
      summary: Add a table to an area (respects area limits)
      parameters:
        - in: path
          name: areaId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTableRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Table" }
        "400":
          description: Invalid request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Area table limit exceeded / VIP max exceeded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Not seeded (run POST /seed first)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /tables:
    get:
      summary: List tables (optionally filter by areaId)
      parameters:
        - in: query
          name: areaId
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Tables
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Table" }
        "503":
          description: Not seeded (run POST /seed first)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /availability:
    get:
      summary: Check availability by date/time/party size with slot-based rules
      parameters:
        - in: query
          name: date
          required: true
          schema: { type: string, format: date, example: "2026-02-18" }
        - in: query
          name: startTime
          required: true
          schema: { type: string, pattern: "^([01]\\d|2[0-3]):[0-5]\\d$", example: "20:00" }
        - in: query
          name: partySize
          required: true
          schema: { type: integer, minimum: 1, example: 7 }
        - in: query
          name: durationMin
          required: false
          schema:
            type: integer
            enum: [90, 120, 180]
            default: 90
        - in: query
          name: areaPreference
          required: false
          schema:
            $ref: "#/components/schemas/AreaPreference"
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AvailabilityResponse" }
        "400":
          description: Invalid query
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Not satisfiable (outside rules/time window)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Not seeded (run POST /seed first)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /reservations:
    post:
      summary: Create reservation (blocks slots from Pending)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReservationRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reservation" }
        "400":
          description: Invalid request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Slot conflict / area limits / overlapping reservation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Not satisfiable (no availability / outside booking window / outside operating hours)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Not seeded (run POST /seed first)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    get:
      summary: List reservations by date (and optional areaId)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: date
          required: true
          schema: { type: string, format: date }
        - in: query
          name: areaId
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Reservations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Reservation" }
        "503":
          description: Not seeded (run POST /seed first)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /reservations/{id}/status:
    patch:
      summary: Update reservation status (confirm/cancel/noshow/complete)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateStatusRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reservation" }
        "400":
          description: Invalid request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Invalid transition / cancel too late not allowed (if you ever choose option A later)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Not seeded (run POST /seed first)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AreaPreference:
      type: string
      enum: [VIP, TERRACE, PATIO, LOBBY, BAR, ANY]
      default: ANY

    Area:
      type: object
      required: [id, code, name, maxTables]
      properties:
        id: { type: string, example: "area_terrace" }
        code: { type: string, example: "TERRACE" }
        name: { type: string, example: "Terraza" }
        maxTables: { type: integer, example: 8 }

    Table:
      type: object
      required: [id, areaId, code, capacity, tableType]
      properties:
        id: { type: string, example: "tbl_terrace_4_01" }
        areaId: { type: string, example: "area_terrace" }
        code: { type: string, example: "T4-01" }
        capacity: { type: integer, example: 4 }
        tableType:
          type: string
          enum: [STANDARD, VIP_ROUND, VIP_SQUARE]
        isActive: { type: boolean, default: true }

    CreateTableRequest:
      type: object
      required: [capacity, tableType]
      properties:
        capacity:
          type: integer
          enum: [2,4,6,8,10]
        tableType:
          type: string
          enum: [STANDARD, VIP_ROUND, VIP_SQUARE]
        code:
          type: string
          description: Optional custom code, else auto-generated.

    ReservationStatus:
      type: string
      enum: [PENDING, CONFIRMED, CANCELLED, NOSHOW, COMPLETED, EXPIRED]

    ReservationSource:
      type: string
      enum: [WHATSAPP, INSTAGRAM, WALKIN, WEB]

    Occasion:
      type: string
      enum: [NONE, BIRTHDAY, ANNIVERSARY, OTHER]

    CreateReservationRequest:
      type: object
      required: [date, startTime, partySize, customerName, phone, source]
      properties:
        date: { type: string, format: date, example: "2026-02-18" }
        startTime: { type: string, example: "20:00" }
        durationMin:
          type: integer
          enum: [90,120,180]
          default: 90
        partySize: { type: integer, minimum: 1, example: 7 }
        areaPreference: { $ref: "#/components/schemas/AreaPreference" }
        vipRequested: { type: boolean, default: false }
        customerName: { type: string, example: "Ana Pérez" }
        phone: { type: string, example: "+1-809-555-1212" }
        email: { type: string, format: email, nullable: true }
        notes: { type: string, nullable: true }
        occasion: { $ref: "#/components/schemas/Occasion" }
        source: { $ref: "#/components/schemas/ReservationSource" }
        assignedTableId:
          type: string
          nullable: true
          description: Optional manual assignment. If provided, system validates slots.

    Reservation:
      type: object
      required:
        [id, status, date, startTime, durationMin, partySize, requiredCapacity,
         areaAssigned, vipRequested, vipFulfilled, estimatedSpend, createdAt]
      properties:
        id: { type: string, example: "rsv_01HR..." }
        status: { $ref: "#/components/schemas/ReservationStatus" }
        date: { type: string, format: date }
        startTime: { type: string, example: "20:00" }
        durationMin: { type: integer, example: 90 }
        partySize: { type: integer, example: 7 }
        requiredCapacity:
          type: integer
          description: Rounded up capacity (e.g., 7->8)
          example: 8
        areaAssigned:
          type: string
          enum: [VIP, TERRACE, PATIO, LOBBY, BAR]
          example: "VIP"
        tableAssignedIds:
          type: array
          items: { type: string }
          description: For VIP A+B, returns [A,B]. Otherwise [single].
        vipRequested: { type: boolean, example: true }
        vipFulfilled: { type: boolean, example: false }
        customerName: { type: string }
        phone: { type: string }
        email: { type: string, format: email, nullable: true }
        notes: { type: string, nullable: true }
        occasion: { $ref: "#/components/schemas/Occasion" }
        source: { $ref: "#/components/schemas/ReservationSource" }
        estimatedSpend:
          type: integer
          description: Calculated as partySize * 700 (RD$)
          example: 4900
        cancelPolicy:
          type: object
          properties:
            freeCancelBeforeMinutes: { type: integer, example: 120 }
            lateCancelPenaltyRate: { type: number, example: 0.5 }
            perPersonEstimate: { type: integer, example: 700 }
        penalty:
          type: object
          nullable: true
          properties:
            applied: { type: boolean, example: true }
            amount: { type: integer, example: 2450 }
            currency: { type: string, example: "DOP" }
            reason: { type: string, example: "Late cancellation (< 2h)" }
        createdAt: { type: string, example: "2026-02-18T19:00:00-04:00" }
        expiresAt:
          type: string
          nullable: true
          description: Pending expiration (createdAt + 15m) when status=PENDING.

    UpdateStatusRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/ReservationStatus" }
        cancelledAt:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true

    AvailabilityResponse:
      type: object
      required: [date, startTime, durationMin, partySize, requiredCapacity, operatingHours, bookingWindow, areaPriority, options]
      properties:
        date: { type: string, format: date }
        startTime: { type: string }
        durationMin: { type: integer }
        partySize: { type: integer }
        requiredCapacity: { type: integer }
        operatingHours:
          type: object
          properties:
            opensAt: { type: string, example: "16:00" }
            closesAt: { type: string, example: "00:00" }
            timezone: { type: string, example: "America/Santo_Domingo" }
        bookingWindow:
          type: object
          properties:
            minLeadMinutes: { type: integer, example: 45 }
            maxDaysAhead: { type: integer, example: 4 }
        areaPriority:
          type: array
          items: { type: string }
          example: ["VIP","TERRACE","PATIO","LOBBY","BAR"]
        options:
          type: array
          items:
            type: object
            required: [area, feasible, candidates]
            properties:
              area: { type: string, example: "VIP" }
              feasible: { type: boolean, example: true }
              candidates:
                type: array
                items:
                  type: object
                  properties:
                    tableIds:
                      type: array
                      items: { type: string }
                      example: ["vip_a","vip_b"]
                    capacity: { type: integer, example: 6 }
                    note: { type: string, example: "VIP combo A+B" }

    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: "NO_AVAILABILITY" }
        message: { type: string, example: "No available tables for 2026-02-18 20:00." }
        details:
          type: object
          additionalProperties: true